requirements.txt
Django==2.2.6
djangorestframework==3.10.3
loginform==1.2.0
lxml==4.4.1
pytz==2019.3
sqlparse==0.3.0
